backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
pydantic-settings>=2.6.0
openai>=1.55.0
python-dotenv>=1.0.1
pillow>=10.4.0
numpy>=1.26.0
opencv-python-headless>=4.10.0
# Legacy Face Mesh (mp.solutions) — newer 0.10.31+ removed `solutions`
mediapipe>=0.10.14,<0.10.30
replicate>=0.34.0
httpx>=0.27.0
